requirements.txt
torch~=2.10.0

pillow~=12.1.1
transformers~=5.1.0
numpy~=1.26.4

PyYAML~=6.0.3
requests~=2.32.5

timm~=1.0.24

torchvision~=0.25.0
